# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/incoterms:
    get:
      tags:
        - Incoterms
      summary: List Incoterms
      description: Get a paginated list of Incoterms with optional filtering and sorting
      operationId: listIncoterms
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncotermList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Incoterms
      summary: Create Incoterm
      description: Create a new Incoterm
      operationId: createIncoterm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncotermRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incoterm'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/terms_and_conditionses:
    get:
      tags:
        - TermsAndConditionses
      summary: List TermsAndConditionses
      description: Get a paginated list of TermsAndConditionses with optional filtering and sorting
      operationId: listTermsAndConditionses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermsAndConditionsList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - TermsAndConditionses
      summary: Create TermsAndConditions
      description: Create a new TermsAndConditions
      operationId: createTermsAndConditions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTermsAndConditionsRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TermsAndConditions'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    Incoterm:
      type: object
      required:
        - id
        - code
        - name
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 8
        name:
          type: string
          maxLength: 120
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreateIncotermRequest:
      type: object
      required:
        - code
        - name
        - is_active
        - metadata
      properties:
        code:
          type: string
          maxLength: 8
        name:
          type: string
          maxLength: 120
        is_active:
          type: boolean
        metadata:
          type: object
...
    IncotermList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Incoterm'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    TermsAndConditions:
      type: object
      required:
        - id
        - code
        - title
        - body
        - is_active
        - metadata
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        title:
          type: string
          maxLength: 150
        body:
          type: string
          maxLength: 8000
        is_active:
          type: boolean
        metadata:
          type: object
...
    CreateTermsAndConditionsRequest:
      type: object
      required:
        - code
        - title
        - body
        - is_active
        - metadata
      properties:
        code:
          type: string
          maxLength: 40
        title:
          type: string
          maxLength: 150
        body:
          type: string
          maxLength: 8000
        is_active:
          type: boolean
        metadata:
          type: object
...
    TermsAndConditionsList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TermsAndConditions'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []